Move gradient to class
